requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
pyarrow==14.0.1
tabulate==0.9.0
tqdm==4.66.1
psutil==5.9.7
colorama==0.4.6
loguru==0.7.2
pyorc==0.8.0
streamlit==1.29.0
plotly==5.18.0 